Changed data type to work

+++ src/main/java/nl/fontys/s3/business/DeleteChatbotFAQ.java
@@ -1,5 +1,5 @@
 package nl.fontys.s3.business;
 
 public interface DeleteChatbotFAQ {
-    void deleteFAQ(int FAQId);
+    void deleteFAQ(Long FAQId);
 }
